pypackages/mock_sensor/pyproject.toml
[project]
name = "mock-sensor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Yin-Chi Chan", email = "ycc39@cam.ac.uk" }
]
requires-python = "==3.13.*"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"
